dev-requirements.txt
tox
mock
coverage==3.7.1
vcrpy
